Move JsonSerializer and XmlSerializer into implementation
