+++ sdk/clientcore/core/src/main/java/io/clientcore/core/implementation/util/XmlSerializer.java
@@ -1,13 +1,15 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-package io.clientcore.core.util.serializer;
+package io.clientcore.core.implementation.util;
 
 import io.clientcore.core.implementation.TypeUtil;
 import io.clientcore.core.serialization.xml.XmlReader;
 import io.clientcore.core.serialization.xml.XmlSerializable;
 import io.clientcore.core.serialization.xml.XmlWriter;
 import io.clientcore.core.util.ClientLogger;
+import io.clientcore.core.util.serializer.ObjectSerializer;
+import io.clientcore.core.util.serializer.SerializationFormat;
 
 import javax.xml.stream.XMLStreamException;
 import java.io.ByteArrayOutputStream;
