+++ sdk/clientcore/core/src/main/java/io/clientcore/core/util/binarydata/BinaryData.java
@@ -4,7 +4,7 @@
 package io.clientcore.core.util.binarydata;
 
 import io.clientcore.core.serialization.json.JsonWriter;
-import io.clientcore.core.util.serializer.JsonSerializer;
+import io.clientcore.core.implementation.util.JsonSerializer;
 import io.clientcore.core.util.serializer.ObjectSerializer;
 
 import java.io.Closeable;
