+++ sdk/clientcore/core/src/test/java/io/clientcore/core/http/MockHttpResponse.java
@@ -7,7 +7,7 @@ import io.clientcore.core.http.models.HttpHeaders;
 import io.clientcore.core.http.models.HttpRequest;
 import io.clientcore.core.http.models.HttpResponse;
 import io.clientcore.core.util.binarydata.BinaryData;
-import io.clientcore.core.util.serializer.JsonSerializer;
+import io.clientcore.core.implementation.util.JsonSerializer;
 import io.clientcore.core.util.serializer.ObjectSerializer;
 
 public class MockHttpResponse extends HttpResponse<BinaryData> {
