+++ sdk/clientcore/core/src/test/java/io/clientcore/core/util/serializer/MockSerializer.java
@@ -3,6 +3,8 @@
 
 package io.clientcore.core.util.serializer;
 
+import io.clientcore.core.implementation.util.JsonSerializer;
+
 import java.io.InputStream;
 import java.io.OutputStream;
 import java.lang.reflect.Type;
